+++ internal/domain/mutagens/arithmetic.go
@@ -1,3 +1,4 @@
+// Package mutagens provides mutation generators for different mutation types.
 package mutagens
 
 import (
@@ -7,6 +8,7 @@ import (
 	m "github.com/mouse-blink/gooze/internal/model"
 )
 
+// GenerateArithmeticMutations generates arithmetic mutations for the given AST node.
 func GenerateArithmeticMutations(n ast.Node, fset *token.FileSet, content []byte, source m.Source, mutationID *int) []m.Mutation {
 	binExpr, ok := n.(*ast.BinaryExpr)
 	if !ok {
@@ -26,11 +28,12 @@ func GenerateArithmeticMutations(n ast.Node, fset *token.FileSet, content []byte
 	end := start + len(original)
 
 	var mutations []m.Mutation
+
 	for _, mutatedOp := range getArithmeticAlternatives(binExpr.Op) {
 		*mutationID++
 		mutatedCode := replaceRange(content, start, end, mutatedOp.String())
 		mutations = append(mutations, m.Mutation{
-			ID:          uint(*mutationID - 1),
+			ID:          *mutationID - 1,
 			Source:      source,
 			Type:        m.MutationArithmetic,
 			MutatedCode: mutatedCode,
@@ -48,6 +51,7 @@ func getArithmeticAlternatives(original token.Token) []token.Token {
 	allOps := []token.Token{token.ADD, token.SUB, token.MUL, token.QUO, token.REM}
 
 	var alternatives []token.Token
+
 	for _, op := range allOps {
 		if op != original {
 			alternatives = append(alternatives, op)
